makefile

+++ libft/libft.h
@@ -30,7 +30,7 @@ int		ft_isascii(int c);
 void	*ft_calloc(size_t num, size_t size);
 size_t	ft_strlen(const char *str);
 int 	hex_len(unsigned long long nbr);
-int 	hex_len_ui(unsigned int nbr);
+int 	hex_len_ui(unsigned long long nbr);
 void	*ft_memset(void *dest, int val, size_t len);
 void	ft_bzero(void *s, size_t n);
 void	*ft_memcpy(void *dst, const void *src, size_t n);
